requirements.txt
telethon
aiohttp
psycopg2-binary
python-dotenv